requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
scikit-learn==1.4.2
sentence-transformers==2.6.1
faiss-cpu==1.8.0
requests
python-dotenv
pycryptodome==3.20.0